moved subscribe event engine initialisation to subscribeOpertaion
